Crate wabt

Bindings to the wabt library.

Modules

script
Module for parsing WebAssembly script format (a.k.a. wast).

Structs

Error
A structure to represent errors coming out from wabt.

Features
Represents which WebAssembly features are enabled in Wabt.

Module
WebAssembly module.

ReadBinaryOptions
Options for reading read binary.

WabtBuf
Buffer returned by wabt.

Wasm2Wat
A builder for converting wasm binary to wasm text format.

Wat2Wasm
A builder for translating wasm text source to wasm binary format.

Enums

ErrorKind
ErrorKind describes an error condition from a wasm module operation, as well as the corresponding error message from wabt, if any.

Functions

wasm2wat
Disassemble wasm binary to wasm text format.

wasm2wat_with_features
Disassemble wasm binary to wasm text format with the given features.

wat2wasm
Translate wasm text source to wasm binary format.

wat2wasm_with_features
Translate wasm text source to wasm binary format.