Crate wabt

Bindings to the wabt library.

Modules

script
Module for parsing WebAssembly script format (a.k.a. wast).

Structs

Error
A structure to represent errors coming out from wabt.

Module
WebAssembly module.

ReadBinaryOptions
Options for reading read binary.

WabtBuf
Buffer returned by wabt.

Wasm2Wat
A builder for converting wasm binary to wasm text format.

Wat2Wasm
A builder for translate wasm text source to wasm binary format.

Functions

wasm2wat
Disassemble wasm binary to wasm text format.

wat2wasm
Translate wasm text source to wasm binary format.