Enum wabt::script::CommandKind

pub enum CommandKind<F32 = f32, F64 = f64> {
    Module {
        module: ModuleBinary,
        name: Option<String>,
    },
    AssertReturn {
        action: Action<F32, F64>,
        expected: Vec<Value<F32, F64>>,
    },
    AssertReturnCanonicalNan {
        action: Action<F32, F64>,
    },
    AssertReturnArithmeticNan {
        action: Action<F32, F64>,
    },
    AssertTrap {
        action: Action<F32, F64>,
        message: String,
    },
    AssertInvalid {
        module: ModuleBinary,
        message: String,
    },
    AssertMalformed {
        module: ModuleBinary,
        message: String,
    },
    AssertUninstantiable {
        module: ModuleBinary,
        message: String,
    },
    AssertExhaustion {
        action: Action<F32, F64>,
    },
    AssertUnlinkable {
        module: ModuleBinary,
        message: String,
    },
    Register {
        name: Option<String>,
        as_name: String,
    },
    PerformAction(Action<F32, F64>),
}

Script's command.

Variants
Module

Define, validate and instantiate a module.

Fields of Module

module: ModuleBinary

Wasm module binary to define, validate and instantiate.

name: Option<String>

If the name is specified, the module should be registered under this name.

AssertReturn

Assert that specified action should yield specified results.

Fields of AssertReturn

action: Action<F32, F64>

Action to perform.

expected: Vec<Value<F32, F64>>

Values that expected to be yielded by the action.

AssertReturnCanonicalNan

Assert that specified action should yield NaN in canonical form.

Fields of AssertReturnCanonicalNan

action: Action<F32, F64>

Action to perform.

AssertReturnArithmeticNan

Assert that specified action should yield NaN with 1 in MSB of fraction field.

Fields of AssertReturnArithmeticNan

action: Action<F32, F64>

Action to perform.

AssertTrap

Assert that performing specified action must yield in a trap.

Fields of AssertTrap

action: Action<F32, F64>

Action to perform.

message: String

Expected failure should be with this message.

AssertInvalid

Assert that specified module is invalid.

Fields of AssertInvalid

module: ModuleBinary

Module that should be invalid.

message: String

Expected failure should be with this message.

AssertMalformed

Assert that specified module cannot be decoded.

Fields of AssertMalformed

module: ModuleBinary

Module that should be malformed.

message: String

Expected failure should be with this message.

AssertUninstantiable

Assert that specified module is uninstantiable.

Fields of AssertUninstantiable

module: ModuleBinary

Module that should be uninstantiable.

message: String

Expected failure should be with this message.

AssertExhaustion

Assert that specified action should yield in resource exhaustion.

Fields of AssertExhaustion

action: Action<F32, F64>

Action to perform.

AssertUnlinkable

Assert that specified module fails to link.

Fields of AssertUnlinkable

module: ModuleBinary

Module that should be unlinkable.

message: String

Expected failure should be with this message.

Register

Register a module under specified name (as_name).

Fields of Register

name: Option<String>

Name of the module, which should be registered under different name.

If None then the last defined module should be used.

as_name: String

New name of the specified module.

PerformAction(Action<F32, F64>)

Perform the specified action.